backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==4.4.1
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
